refactor: 배달 요금 double -> BigDecimal 변경

+++ appstore-core/domain/src/main/java/com/example/delivery/model/dto/response/DeliveryDetailResponse.java
@@ -11,6 +11,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -37,9 +38,9 @@ public class DeliveryDetailResponse {
 
     private String deliveryOptions;
 
-    private double deliveryFee; //BigDemical
+    private BigDecimal deliveryFee; //BigDemical
 
-    private DeliveryDetailResponse(String reservationNumber, PersonalDeliveryCategory deliveryPersonalCategory, LocalDateTime deliveryDate, DeliveryAddress deliveryAddress, List<DeliveryAddress> stopOverAddresses, Vehicle vehicle, String deliveryOptions, double deliveryFee) {
+    private DeliveryDetailResponse(String reservationNumber, PersonalDeliveryCategory deliveryPersonalCategory, LocalDateTime deliveryDate, DeliveryAddress deliveryAddress, List<DeliveryAddress> stopOverAddresses, Vehicle vehicle, String deliveryOptions, BigDecimal deliveryFee) {
         this.reservationNumber = reservationNumber;
         this.deliveryPersonalCategory = deliveryPersonalCategory;
         this.deliveryDate = deliveryDate;
